added series update
